Reorganize into class namespaces and use phc.easy as entrypoint
