+++ phc/easy/patients/__init__.py
@@ -2,7 +2,7 @@ import os
 
 import pandas as pd
 from phc.easy.auth import Auth
-from phc.easy import expand_data_frame
+from phc.easy.frame import Frame
 from phc.easy.patients.address import expand_address_column
 from phc.easy.patients.name import expand_name_column
 from phc.services import Fhir
@@ -19,14 +19,12 @@ class Patient:
         if actual_count > limit:
             print(f'Retrieved {limit}/{actual_count} patients')
 
-        df = pd.DataFrame([
-            hit['_source'] for hit in response.data.get('hits').get('hits')
-        ])
+        df = pd.DataFrame(
+            [hit['_source'] for hit in response.data.get('hits').get('hits')])
 
         if raw:
             return df
 
-        return expand_data_frame(df,
-                                 custom_columns=[('address',
-                                                  expand_address_column),
-                                                 ('name', expand_name_column)])
+        return Frame.expand(df,
+                            custom_columns=[('address', expand_address_column),
+                                            ('name', expand_name_column)])
